refactor: change components directory layout
