+++ src/components/Application/types.ts
@@ -1,0 +1,9 @@
+interface IApplicationProps {
+    node: ({ }: { appId: string }) => JSX.Element;
+    title: string;
+    id: string;
+    x?: number;
+    y?: number;
+}
+
+export type { IApplicationProps };
