+++ src/components/Calculator/styles.ts
