+++ src/components/Calculator/types.ts
@@ -1,0 +1,7 @@
+interface ICalculatorProps {
+    text: string;
+    onClick?: (value: string) => void;
+    className?: string;
+}
+
+export type { ICalculatorProps };
