+++ src/components/Card/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+interface ICardProps extends React.HTMLAttributes<HTMLDivElement> {
+    children: ReactNode;
+    isDraggable?: boolean;
+    width?: number;
+    height?: number;
+    defaultPosition?: { x: number; y: number };
+}
+
+export type { ICardProps };
