+++ src/components/Dropdown/types.ts
@@ -1,0 +1,11 @@
+import { ReactElement } from "react";
+
+interface IDropdownProps {
+    trigger: ReactElement;
+    items: {
+        id: string;
+        node: ReactElement;
+    }[];
+};
+
+export type { IDropdownProps };
