+++ src/components/Terminal/types.ts
@@ -1,0 +1,11 @@
+interface Commands {
+    [key: string]: TerminalCommand;
+}
+
+interface TerminalCommand {
+    description: string;
+    usage: string;
+    fn: (...args: string[]) => string | Promise<string>;
+}
+
+export type { Commands }
