+++ src/components/WellcomeCard/styles.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+const Content = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  gap: 8px;
+
+  width: 100%;
+  height: 100%;
+`;
+
+export { Content };
