+++ src/pages/Main.tsx
@@ -1,8 +1,8 @@
-import { lazy, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 
-import Loading from "./Loading";
-import Desktop from "./Desktop";
 import styled from "styled-components";
+import Desktop from "./Desktop";
+import Loading from "./Loading";
 
 const Container = styled.div`
   all: unset;
